Move semantic token types to a separate file
To declutter the entry point
