+++ src/extension.ts
@@ -1,54 +1,8 @@
 import * as vscode from 'vscode';
 import { spawnSync } from 'node:child_process';
+import { type TokenModifier, type TokenType, legend } from './semanticTokens';
 
 // Inspired by https://github.com/microsoft/vscode-extension-samples/tree/main/semantic-tokens-sample
-
-// https://code.visualstudio.com/api/language-extensions/semantic-highlight-guide#standard-token-types-and-modifiers
-const tokenTypes = [
-  'namespace',
-  'class',
-  'enum',
-  'interface',
-  'struct',
-  'typeParameter',
-  'type',
-  'parameter',
-  'variable',
-  'property',
-  'enumMember',
-  'decorator',
-  'event',
-  'function',
-  'method',
-  'macro',
-  'label',
-  'comment',
-  'string',
-  'keyword',
-  'number',
-  'regexp',
-  'operator',
-] as const;
-
-const tokenModifiers = [
-  'declaration',
-  'definition',
-  'readonly',
-  'static',
-  'deprecated',
-  'abstract',
-  'async',
-  'modification',
-  'documentation',
-  'defaultLibrary',
-] as const;
-
-// A workaround because a readonly string[] is not assignable to a normal string[]
-type Writeable<T> = { -readonly [P in keyof T]: T[P] };
-const legend = new vscode.SemanticTokensLegend(
-  tokenTypes as Writeable<typeof tokenTypes>,
-  tokenModifiers as Writeable<typeof tokenModifiers>
-);
 
 let output = vscode.window.createOutputChannel('Rzk');
 export function activate(context: vscode.ExtensionContext) {
@@ -61,9 +15,6 @@ export function activate(context: vscode.ExtensionContext) {
     )
   );
 }
-
-type TokenType = (typeof tokenTypes)[number];
-type TokenModifier = (typeof tokenModifiers)[number];
 
 interface ParsedToken {
   line: number;
